Remove unecessary casts bug 35836, from Nathan Beyer
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/primitives/trunk@278573 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/collections/primitives/ArrayUnsignedByteList.java
@@ -233,7 +233,7 @@ public class ArrayUnsignedByteList extends RandomAccessShortList implements Shor
     //-------------------------------------------------------------------------
 
     private final short toShort(byte value) { 
-        return (short)(((short)value)&MAX_VALUE);
+        return (short)(value & MAX_VALUE);
     }
 
     private final byte fromShort(short value) {
